Extract private FFI type defintions to separate header ffi-priv-types.h
